Update config loading and config value retrieval code.
We now use configobj library plus additional abstractions for retrieving
values which make code more robust and easier to change the underlying
implementation later on.

In addition to that, we now support updating the config (e.g. if it was
changed by the admin plugin) and re-loading it if the file has been
modified since the last load time.